requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
alembic
python-dotenv
passlib[bcrypt]
pydantic
python-jose
jinja2